requirements.txt
langchain>=0.3.13
langchain-core>=0.3.28
langchain-community>=0.3.13
langchain-openai>=0.2.14
langgraph>=1.0.4
openai>=1.58.1
python-dotenv>=1.2.1
flask>=3.1.0
flask-cors>=5.0.0
